Rearrange where tests are located to be more logical
