+++ .codecov.yml
@@ -16,6 +16,9 @@ coverage:
         if_not_found: success
         if_ci_failed: error
     project: *cov
+    ignore:
+      - "src/__testUtils__/"
+      - "__testUtils__"
 comment:
   layout: "reach, diff, flags, files"
   behavior: default
